renamed rocket_release illustration to match the template name
