+++ webapp/src/components/assets/illustrations/product_release_svg.tsx
